requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
qdrant-client==1.6.1
google-generativeai==0.5.2
python-dotenv==1.0.1